- Renamed pseudoinverse_calculation -> inverse_jacobian_calculation
